Add two possibilities for calculating payments
One actually tracks pairs of payments, the other uses the pools and just
figures out who should right a cheque to whom

The advantage of pairs is that an unlucky creditor doesn't get screwed
by being paired with a deadbeat debtor.

The advantage of pools is that there are far fewer payments to make.